Feat: User-Interest Entity 생성

+++ backend/src/main/java/com/orleave/entity/MeetingSetting.java
@@ -19,7 +19,7 @@ public class MeetingSetting {
 	
 	@Id
 	@Column(name = "user_no", nullable = false, columnDefinition = "INT UNSIGNED")
-	private int userNo;
+	private Long userNo;
 	
 	@OneToOne
 	@MapsId
